Warnings.

+++ src/modules/mime/mimetic/rfc822/group.cxx
@@ -45,7 +45,7 @@ static string::size_type find_not_in_quote(const string& s, const string::value_
 {
     size_t len = s.length();
     bool in_dquote = false;
-    for(int i =0; i < len; ++i)
+    for(int i =0; i < (int) len; ++i)
     {
         if(s[i] == '"')
             in_dquote = !in_dquote;
